requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.0.0
urllib3==1.26.7
zipp==3.6.0
